+++ tests/fixtures/source/InvalidLogicalOperands.xs
@@ -1,0 +1,10 @@
+// SPDX-FileCopyrightText: 2026 Leitwolf <xs-lang.chess031@slmails.com>
+// SPDX-License-Identifier: Apache-2.0
+
+fn main() -> Long {
+  invalid: Bool = 1 && true;
+  if (invalid) {
+    return 1;
+  }
+  return 0;
+}