+++ xslang/sources/hir/text/parser.rs
@@ -654,6 +654,8 @@ impl Parser<'_>
       "bit_and" => Some(BinaryOperator::BitAnd),
       "bit_or" => Some(BinaryOperator::BitOr),
       "bit_xor" => Some(BinaryOperator::BitXor),
+      "logical_and" => Some(BinaryOperator::LogicalAnd),
+      "logical_or" => Some(BinaryOperator::LogicalOr),
       "shift_left" => Some(BinaryOperator::ShiftLeft),
       "shift_right" => Some(BinaryOperator::ShiftRight),
       "eq" => Some(BinaryOperator::Equal),